finally catcher is now correctly moving and catching hit objects, in osu mania improved seeking, in catch all spinners now should calculate amount of bananas correctly to correctly update RNG values for slider droplets

+++ OsuFileParsers/Classes/Beatmap/osu/Objects/CatchBananaShowerData.cs
@@ -4,6 +4,6 @@ namespace OsuFileParsers.Classes.Beatmap.osu.Objects
 {
     public class CatchBananaShowerData : HitObjectData
     {
-        public int EndTime { get; set; }
+        public double EndTime { get; set; }
     }
 }
